+++ src/components/recipeDetail/IngredientSearch.tsx
@@ -45,9 +45,8 @@ export const IngredientSearch: FC<{ recipe: Recipe }> = ({ recipe }) => {
       }}
       filterOptions={(options, params) => {
         const filtered = filter(options, params);
-
         const { inputValue } = params;
-        // Suggest the creation of a new value
+
         const isExisting = options.some((option) => inputValue === option.name);
         if (inputValue !== "" && !isExisting) {
           filtered.push({
